app/app_bis/requirements.txt
numpy==1.26.4
opencv-python==4.7.0.72
dotenv
pillow
onnxruntime
torchvision
langchain_core
langchain_mistralai
streamlit
streamlit-webrtc